+++ src/styles/GlobalStyle.ts
@@ -1,6 +1,6 @@
 import { createGlobalStyle } from 'styled-components'
 
-export const GlobalStyles = createGlobalStyle`
+const GlobalStyle = createGlobalStyle`
   html,
   body {
     padding: 0;
@@ -60,3 +60,5 @@ export const GlobalStyles = createGlobalStyle`
     height: 100%;
   }
 `
+
+export default GlobalStyle
